split reactive constructor
