--- event-store/impl-postgresql/src/commonMain/kotlin/dev/eskt/store/impl/pg/PostgresqlEventStore.kt
@@ -1,25 +1,0 @@
-package dev.eskt.store.impl.pg
-
-@Suppress("FunctionName")
-public fun PostgresqlEventStore(
-    dataSource: DataSource,
-    eventTable: String,
-    eventWriteTable: String = eventTable,
-    block: PostgresqlConfigBuilder<DataSource>.() -> Unit,
-): PostgresqlJdbcEventStore = PostgresqlJdbcEventStore(
-    PostgresqlConfigBuilder(dataSource, eventTable, eventWriteTable)
-        .apply(block)
-        .build(),
-)
-
-@Suppress("FunctionName")
-public fun PostgresqlEventStore(
-    dataSource: ReactiveDataSource,
-    eventTable: String,
-    eventWriteTable: String = eventTable,
-    block: PostgresqlConfigBuilder<ReactiveDataSource>.() -> Unit,
-): PostgresqlR2dbcEventStore = PostgresqlR2dbcEventStore(
-    PostgresqlConfigBuilder(dataSource, eventTable, eventWriteTable)
-        .apply(block)
-        .build(),
-)
