+++ event-store/impl-postgresql/src/commonMain/kotlin/dev/eskt/store/impl/pg/data-source-blocking.kt
@@ -3,3 +3,15 @@ package dev.eskt.store.impl.pg
 public expect interface DataSource
 
 internal expect fun DataSource.createAdapter(): dev.eskt.store.impl.pg.blocking.DatabaseAdapter
+
+@Suppress("FunctionName")
+public fun PostgresqlEventStore(
+    dataSource: DataSource,
+    eventTable: String,
+    eventWriteTable: String = eventTable,
+    block: PostgresqlConfigBuilder<DataSource>.() -> Unit,
+): PostgresqlJdbcEventStore = PostgresqlJdbcEventStore(
+    PostgresqlConfigBuilder(dataSource, eventTable, eventWriteTable)
+        .apply(block)
+        .build(),
+)
